+++ R/geodist.R
@@ -12,6 +12,9 @@
 #' @param role or model term created by this step, what analysis
 #'  role should be assigned?. By default, the function assumes
 #'  that resulting distance will be used as a predictor in a model.
+#' @param is_lat_lon A logical: Are coordinates in latitude and longitude? If
+#'  TRUE the Haversine formula is used and the returned result is meters. If
+#'  FALSE the Pythagorean formula is used. Default is TRUE.
 #' @param log A logical: should the distance be transformed by
 #'  the natural log function?
 #' @param columns A character string of variable names that will
@@ -22,11 +25,14 @@
 #' @return An updated version of `recipe` with the new step added
 #'  to the sequence of existing steps (if any).
 #' @details When you [`tidy()`] this step, a tibble with columns echoing the
-#'  values of `lat`, `lon`, `ref_lat`, `ref_lon`, `name`, and `id` is returned.
+#'  values of `lat`, `lon`, `ref_lat`, `ref_lon`, `is_lat_lon`, `name`, and `id`
+#'  is returned.
 #' @keywords datagen
 #' @concept preprocessing
 #' @export
-#' @details `step_geodist` will create a
+#' @details `step_geodist` uses the Pythagorean theorem to calculate Euclidean
+#'  distances if `is_lat_lon` is FALSE. If `is_lat_lon` is TRUE, the Haversine
+#'  formula is used to calculate the great-circle distance in meters.
 #'
 #' @examples
 #'
@@ -37,7 +43,8 @@
 #' near_station <- recipe( ~ ., data = Smithsonian) %>%
 #'   update_role(name, new_role = "location") %>%
 #'   step_geodist(lat = latitude, lon = longitude, log = FALSE,
-#'                ref_lat = 38.8986312, ref_lon = -77.0062457) %>%
+#'                ref_lat = 38.8986312, ref_lon = -77.0062457,
+#'                is_lat_lon = TRUE) %>%
 #'   prep(training = Smithsonian)
 #'
 #' bake(near_station, new_data = NULL) %>%
@@ -51,6 +58,7 @@ step_geodist <- function(recipe,
                          trained = FALSE,
                          ref_lat = NULL,
                          ref_lon = NULL,
+                         is_lat_lon = TRUE,
                          log = FALSE,
                          name = "geo_dist",
                          columns = NULL,
@@ -60,10 +68,18 @@ step_geodist <- function(recipe,
     rlang::abort("`ref_lon` should be a single numeric value.")
   if (length(ref_lat) != 1 || !is.numeric(ref_lat))
     rlang::abort("`ref_lat` should be a single numeric value.")
+  if (length(is_lat_lon) != 1 || !is.logical(is_lat_lon))
+    rlang::abort("`is_lat_lon` should be a single logical value.")
   if (length(log) != 1 || !is.logical(log))
     rlang::abort("`log` should be a single logical value.")
   if (length(name) != 1 || !is.character(name))
     rlang::abort("`name` should be a single character value.")
+
+  if (is_lat_lon && abs(ref_lat) > 90.0)
+    rlang::abort("`ref_lat` should be between -90 and 90")
+  if (is_lat_lon && abs(ref_lon) > 180.0)
+    rlang::abort("`ref_lon` should be between -180 and 180")
+
 
   add_step(
     recipe,
@@ -74,6 +90,7 @@ step_geodist <- function(recipe,
       trained = trained,
       ref_lon = ref_lon,
       ref_lat = ref_lat,
+      is_lat_lon = is_lat_lon,
       log = log,
       name = name,
       columns = columns,
@@ -84,7 +101,7 @@ step_geodist <- function(recipe,
 }
 
 step_geodist_new <-
-  function(lon, lat, role, trained, ref_lon, ref_lat,
+  function(lon, lat, role, trained, ref_lon, ref_lat, is_lat_lon,
            log, name, columns, skip, id) {
     step(
       subclass = "geodist",
@@ -94,6 +111,7 @@ step_geodist_new <-
       trained = trained,
       ref_lon = ref_lon,
       ref_lat = ref_lat,
+      is_lat_lon = is_lat_lon,
       log = log,
       name = name,
       columns = columns,
@@ -108,6 +126,10 @@ prep.step_geodist <- function(x, training, info = NULL, ...) {
   lon_name <- eval_select_recipes(x$lon, training, info)
   lat_name <- eval_select_recipes(x$lat, training, info)
 
+  # for backward compatibility
+  if (!'is_lat_lon' %in% names(x))
+    x$is_lat_lon <- FALSE
+
   if (length(lon_name) > 1)
     rlang::abort("`lon` should resolve to a single column name.")
   check_type(training[, lon_name])
@@ -115,6 +137,7 @@ prep.step_geodist <- function(x, training, info = NULL, ...) {
   if (length(lat_name) > 1)
     rlang::abort("`lat` should resolve to a single column name.")
   check_type(training[, lat_name])
+
 
   if (any(names(training) == x$name))
     rlang::abort("'", x$name, "' is already used in the data.")
@@ -126,6 +149,7 @@ prep.step_geodist <- function(x, training, info = NULL, ...) {
     trained = TRUE,
     ref_lon = x$ref_lon,
     ref_lat = x$ref_lat,
+    is_lat_lon = x$is_lat_lon,
     log = x$log,
     name = x$name,
     columns = c(lat_name, lon_name),
@@ -134,19 +158,67 @@ prep.step_geodist <- function(x, training, info = NULL, ...) {
   )
 }
 
-geo_dist_calc <- function(x, a, b)
-  apply(x, 1, function(x, a, b) sqrt((x[1] - a) ^ 2 + (x[2] - b) ^ 2),
-        a = a, b = b)
+geo_dist_calc_xy <- function(x_1, y_1, x_2, y_2)
+  sqrt((x_1 - x_2) ^ 2L + (y_1 - y_2) ^ 2L)
+
+
+# earth_radius = 6371e3 in meters
+geo_dist_calc_lat_lon <- function(x_1, y_1, x_2, y_2, earth_radius = 6371e3) {
+
+  if(any(abs(x_1) > 180.0)) {
+    rlang::abort("All `lon` values should be between -180 and 180")
+  }
+
+  if(any(abs(y_1) > 90.0)) {
+    rlang::abort("All `lat` values should be between -90 and 90")
+  }
+
+  to_rad <- pi / 180.0
+
+  y_1 <- y_1 * to_rad
+  y_2 <- y_2 * to_rad
+
+  x_1 <- x_1 * to_rad
+  x_2 <- x_2 * to_rad
+
+  delta_lat <- (y_2 - y_1) / 2.0
+  delta_lon <- (x_2 - x_1) / 2.0
+
+  a <- sin(delta_lat) ^ 2L + cos(y_1) * cos(y_2) * sin(delta_lon) ^ 2L
+  c <- 2.0 * atan2(sqrt(a), sqrt(1-a))
+
+  earth_radius * c
+
+}
+
 
 #' @export
 bake.step_geodist <- function(object, new_data, ...) {
-  dist_vals <-
-    geo_dist_calc(new_data[, object$columns], object$ref_lat, object$ref_lon)
+
+  # for backward compatibility
+  if (!'is_lat_lon' %in% names(object))
+    object$is_lat_lon <- FALSE
+
+  if(object$is_lat_lon) {
+    dist_vals <-
+      geo_dist_calc_lat_lon(new_data[[object$columns[2]]], # lon
+                            new_data[[object$columns[1]]], # lat
+                            object$ref_lon,
+                            object$ref_lat)
+  } else {
+    dist_vals <-
+      geo_dist_calc_xy(new_data[[object$columns[2]]], # lon
+                       new_data[[object$columns[1]]], # lat
+                       object$ref_lon,
+                       object$ref_lat)
+  }
+
   if (object$log) {
     new_data[, object$name] <- log(dist_vals)
   } else {
     new_data[, object$name] <- dist_vals
   }
+
   new_data
 }
 
@@ -164,12 +236,18 @@ print.step_geodist <-
 #' @param x A `step_geodist` object.
 #' @export
 tidy.step_geodist <- function(x, ...) {
+
+  # for backward compatibility
+  if (!'is_lat_lon' %in% names(object))
+    object$is_lat_lon <- FALSE
+
   if (is_trained(x)) {
     res <- tibble(
       latitude = x$columns[1],
       longitude = x$columns[2],
       ref_latitude = x$ref_lat,
       ref_longitude = x$ref_lon,
+      is_lat_lon = x$is_lat_lon,
       name = x$name
     )
   } else {
@@ -178,6 +256,7 @@ tidy.step_geodist <- function(x, ...) {
       longitude = sel2char(x$lon),
       ref_latitude = x$ref_lat,
       ref_longitude = x$ref_lon,
+      is_lat_lon = x$is_lat_lon,
       name = x$name
     )
   }